requirements.txt
streamlit==1.19.0
pandas==1.4.4
pillow==9.2.0
numpy==1.24.2
matplotlib==3.5.2
seaborn==0.11.2
plotly==5.9.0
scipy==1.9.1
sklearn==1.0.2
xgboost==1.7.4